+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - Health bars only show up after hitting an enemy, and disappear after 5 seconds.
 - You can use the ESC key to close the store as well.
 - Coins now "burst" out from a dying enemy.
+- Removed the "Not Basic" map.
+- Renamed "Basic" to "Testing Map."
 ### Bugfixes
 - Toolbar is no longer briefly visible at the beginning of a game.
 - Bullets now go through coins.